fixes n improvments

+++ Modules/AdminModule/Http/Controllers/Web/Admin/AdminController.php
@@ -677,7 +677,7 @@ class AdminController extends Controller
 
     public function acknowledgeSetupGuideWelcome(): JsonResponse
     {
-        acknowledgeAdminSetupGuideWelcome((int) auth()->id());
+        acknowledgeAdminSetupGuideWelcome(auth()->id());
 
         return response()->json(['ok' => true]);
     }
